moved compilere filename builder to separate function
